+++ src/fft.rs
@@ -168,7 +168,7 @@ pub mod radix2 {
     use enums;
     use ffi;
 
-    pub fn forward(data: &mut [f64], stride: usize, n: usize) -> enums::Value {
+    pub fn forward(data: &mut [f64], stride: u64, n: u64) -> enums::Value {
         enums::Value::from(unsafe {
             sys::gsl_fft_complex_radix2_forward(data.as_mut_ptr(), stride, n)
         })
@@ -176,8 +176,8 @@ pub mod radix2 {
 
     pub fn transform(
         data: &mut [f64],
-        stride: usize,
-        n: usize,
+        stride: u64,
+        n: u64,
         sign: ::FftDirection,
     ) -> enums::Value {
         enums::Value::from(unsafe {
@@ -185,20 +185,20 @@ pub mod radix2 {
         })
     }
 
-    pub fn backward(data: &mut [f64], stride: usize, n: usize) -> enums::Value {
+    pub fn backward(data: &mut [f64], stride: u64, n: u64) -> enums::Value {
         enums::Value::from(unsafe {
             sys::gsl_fft_complex_radix2_backward(data.as_mut_ptr(), stride, n)
         })
     }
 
-    pub fn inverse(data: &mut [f64], stride: usize, n: usize) -> enums::Value {
+    pub fn inverse(data: &mut [f64], stride: u64, n: u64) -> enums::Value {
         enums::Value::from(unsafe {
             sys::gsl_fft_complex_radix2_inverse(data.as_mut_ptr(), stride, n)
         })
     }
 
     /// This is decimation-in-frequency version of the radix-2 FFT function.
-    pub fn dif_forward(data: &mut [f64], stride: usize, n: usize) -> enums::Value {
+    pub fn dif_forward(data: &mut [f64], stride: u64, n: u64) -> enums::Value {
         enums::Value::from(unsafe {
             sys::gsl_fft_complex_radix2_dif_forward(data.as_mut_ptr(), stride, n)
         })
@@ -207,8 +207,8 @@ pub mod radix2 {
     /// This is decimation-in-frequency version of the radix-2 FFT function.
     pub fn dif_transform(
         data: &mut [f64],
-        stride: usize,
-        n: usize,
+        stride: u64,
+        n: u64,
         sign: ::FftDirection,
     ) -> enums::Value {
         enums::Value::from(unsafe {
@@ -217,14 +217,14 @@ pub mod radix2 {
     }
 
     /// This is decimation-in-frequency version of the radix-2 FFT function.
-    pub fn dif_backward(data: &mut [f64], stride: usize, n: usize) -> enums::Value {
+    pub fn dif_backward(data: &mut [f64], stride: u64, n: u64) -> enums::Value {
         enums::Value::from(unsafe {
             sys::gsl_fft_complex_radix2_dif_backward(data.as_mut_ptr(), stride, n)
         })
     }
 
     /// This is decimation-in-frequency version of the radix-2 FFT function.
-    pub fn dif_inverse(data: &mut [f64], stride: usize, n: usize) -> enums::Value {
+    pub fn dif_inverse(data: &mut [f64], stride: u64, n: u64) -> enums::Value {
         enums::Value::from(unsafe {
             sys::gsl_fft_complex_radix2_dif_inverse(data.as_mut_ptr(), stride, n)
         })
@@ -253,8 +253,8 @@ pub mod mixed_radix {
 
     pub fn forward(
         data: &mut [f64],
-        stride: usize,
-        n: usize,
+        stride: u64,
+        n: u64,
         wavetable: &FftComplexWaveTable,
         work: &mut FftComplexWorkspace,
     ) -> enums::Value {
@@ -271,8 +271,8 @@ pub mod mixed_radix {
 
     pub fn transform(
         data: &mut [f64],
-        stride: usize,
-        n: usize,
+        stride: u64,
+        n: u64,
         wavetable: &FftComplexWaveTable,
         work: &mut FftComplexWorkspace,
         sign: ::FftDirection,
@@ -291,8 +291,8 @@ pub mod mixed_radix {
 
     pub fn backward(
         data: &mut [f64],
-        stride: usize,
-        n: usize,
+        stride: u64,
+        n: u64,
         wavetable: &FftComplexWaveTable,
         work: &mut FftComplexWorkspace,
     ) -> enums::Value {
@@ -309,8 +309,8 @@ pub mod mixed_radix {
 
     pub fn inverse(
         data: &mut [f64],
-        stride: usize,
-        n: usize,
+        stride: u64,
+        n: u64,
         wavetable: &FftComplexWaveTable,
         work: &mut FftComplexWorkspace,
     ) -> enums::Value {
@@ -359,7 +359,7 @@ pub mod real_radix2 {
     /// complex[n-1].imag  =   -data[n-1]
     /// Note that the output data can be converted into the full complex sequence using the function gsl_fft_halfcomplex_radix2_unpack described
     /// below.
-    pub fn transform(data: &mut [f64], stride: usize, n: usize) -> enums::Value {
+    pub fn transform(data: &mut [f64], stride: u64, n: u64) -> enums::Value {
         enums::Value::from(unsafe {
             sys::gsl_fft_real_radix2_transform(data.as_mut_ptr(), stride, n)
         })
@@ -367,7 +367,7 @@ pub mod real_radix2 {
 
     /// This function computes the inverse or backwards in-place radix-2 FFT of length n and stride stride on the half-complex sequence data
     /// stored according the output scheme used by gsl_fft_real_radix2. The result is a real array stored in natural order.
-    pub fn inverse(data: &mut [f64], stride: usize, n: usize) -> enums::Value {
+    pub fn inverse(data: &mut [f64], stride: u64, n: u64) -> enums::Value {
         enums::Value::from(unsafe {
             sys::gsl_fft_halfcomplex_radix2_inverse(data.as_mut_ptr(), stride, n)
         })
@@ -375,7 +375,7 @@ pub mod real_radix2 {
 
     /// This function computes the inverse or backwards in-place radix-2 FFT of length n and stride stride on the half-complex sequence data
     /// stored according the output scheme used by gsl_fft_real_radix2. The result is a real array stored in natural order.
-    pub fn backward(data: &mut [f64], stride: usize, n: usize) -> enums::Value {
+    pub fn backward(data: &mut [f64], stride: u64, n: u64) -> enums::Value {
         enums::Value::from(unsafe {
             sys::gsl_fft_halfcomplex_radix2_backward(data.as_mut_ptr(), stride, n)
         })
@@ -414,8 +414,8 @@ pub mod real_radix2 {
     pub fn unpack(
         halfcomplex_coefficient: &mut [f64],
         complex_coefficient: &mut [f64],
-        stride: usize,
-        n: usize,
+        stride: u64,
+        n: u64,
     ) -> enums::Value {
         enums::Value::from(unsafe {
             sys::gsl_fft_halfcomplex_radix2_unpack(
