+++ src/types/chebyshev.rs
@@ -41,7 +41,7 @@ pub struct ChebSeries {
 }
 
 impl ChebSeries {
-    pub fn new(n: usize) -> Option<ChebSeries> {
+    pub fn new(n: u64) -> Option<ChebSeries> {
         let tmp = unsafe { sys::gsl_cheb_alloc(n) };
 
         if tmp.is_null() {
@@ -94,13 +94,13 @@ impl ChebSeries {
     }
 
     /// This function returns the order of Chebyshev series cs.
-    pub fn order(&self) -> usize {
+    pub fn order(&self) -> u64 {
         unsafe { sys::gsl_cheb_order(self.c) }
     }
 
     /// This function returns the size of the Chebyshev coefficient array c[] for the Chebyshev
     /// series cs.
-    pub fn size(&self) -> usize {
+    pub fn size(&self) -> u64 {
         unsafe { sys::gsl_cheb_size(self.c) }
     }
 
@@ -130,7 +130,7 @@ impl ChebSeries {
 
     /// This function evaluates the Chebyshev series cs at a given point x, to (at most) the given
     /// order order.
-    pub fn eval_n(&self, order: usize, x: f64) -> f64 {
+    pub fn eval_n(&self, order: u64, x: f64) -> f64 {
         unsafe { sys::gsl_cheb_eval_n(self.c, order, x) }
     }
 
@@ -139,7 +139,7 @@ impl ChebSeries {
     /// is made from the first neglected term in the series.
     pub fn eval_n_err(
         &self,
-        order: usize,
+        order: u64,
         x: f64,
         result: &mut f64,
         abs_err: &mut f64,
